requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pymysql>=1.1.0
streamlit>=1.28.0
